Initial re-packaging of Chromium for the iMX8
The choice of 101.0.4951.54 is based on discussions in meta-browser[0]
and other sources. Chromium builds are poorly tested on iMX8 devices,
for some unknown reason. Also, meta-imx imx-5.15.71-2.2.2.xml manifest
includes meta-browser e232c2e (which packages 101.0.4951.54), so we trust
 NXP has tested this build.

If this works maybe we could pave the way for better testing and upstream
integration in Debian by building against testing/newer upstream versions.

[0] https://github.com/OSSystems/meta-browser/issues/711

Signed-off-by: Leonardo Held <leonardo.held@toradex.com>